Implement classification of Eclipse core problems
