+++ core/META-INF/MANIFEST.MF
@@ -36,6 +36,7 @@ Export-Package: com.google.gson,
  nl.tudelft.watchdog.core.logic.ui,
  nl.tudelft.watchdog.core.logic.ui.events,
  nl.tudelft.watchdog.core.logic.ui.listeners,
+ nl.tudelft.watchdog.core.logic.ui.listeners.staticanalysis,
  nl.tudelft.watchdog.core.ui.preferences,
  nl.tudelft.watchdog.core.ui.util,
  nl.tudelft.watchdog.core.ui.wizards,
