+++ eclipse/plugin/META-INF/MANIFEST.MF
@@ -19,6 +19,8 @@ Import-Package: junit.runner,
  org.eclipse.jdt.junit,
  org.eclipse.jdt.junit.model,
  org.eclipse.jdt.ui,
+ org.eclipse.jdt.internal.compiler.util,
+ org.eclipse.jdt.internal.compiler.problem,
  org.eclipse.jface.action,
  org.eclipse.jface.dialogs,
  org.eclipse.jface.layout,
